requirements.txt
locust==2.33.0
assertpy==1.1
requests==2.32.3
python-dotenv==1.0.1
